API BREAK: eina_list_search_sorted_near_list() gets one more parameter.
This should not impact anybody, at least in SVN I got no hits for this
function.

The new parameter contains the result of the last call to func(), so
we can know if the node is smaller, bigger or exactly the requested
value and don't need to call func() on node to know for sure.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk/eina@41623 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33