more efficient model properties based on struct.
this is a killer, should be very efficient in memory and speed to
set/get items: instead of a hash of properties, keep them in a C
struct!

The constraint is that properties have fixed types defined at compile
time and cannot be deleted, but this is expected in many cases (ie:
esql rows).




git-svn-id: http://svn.enlightenment.org/svn/e/trunk/eina@67517 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33